refactor(essence): 重构自定义武器弹窗，提取公共表单组件
1.  从custom-weapon-dialog.tsx抽离武器表单为独立的WeaponFormDialog组件
2.  移除settings页面和弹窗中多余的blob URL释放逻辑
3.  为设置页面添加国际化多语言支持
4.  移除未使用的Upload图标导入
